added raw storage and typed storage just in case I need them later

+++ crates/handler/src/lib.rs
@@ -110,11 +110,11 @@ fn handle_glutin_events(sleeper: &mut LoopHelper, world: &mut World, event: Even
 fn handle_device_event(event: DeviceEvent, world: &mut World, _control_flow: &mut ControlFlow) {
     match event {
         DeviceEvent::MouseMotion { delta } => {
-            world.input.receive_mouse_position_event(veclib::vec2(delta.0, delta.1));
+            world.input.receive_mouse_position_event(veclib::vec2(delta.0 as f32, delta.1 as f32));
         }
         DeviceEvent::MouseWheel { delta } => match delta {
-            glutin::event::MouseScrollDelta::LineDelta(_x, y) => world.input.receive_mouse_scroll_event(y as f64),
-            glutin::event::MouseScrollDelta::PixelDelta(y) => world.input.receive_mouse_scroll_event(y.x),
+            glutin::event::MouseScrollDelta::LineDelta(_x, y) => world.input.receive_mouse_scroll_event(y),
+            glutin::event::MouseScrollDelta::PixelDelta(y) => world.input.receive_mouse_scroll_event(y.x as f32),
         },
         DeviceEvent::Key(input) => {
             if let Some(virtual_keycode) = input.virtual_keycode {
